refactor: split Foods into custom hook
